# openapi/raibitserver.yaml
openapi: 3.1.0
info:
  title: RAIBITSERVER API
  version: 0.2.0
  description: Canonical desired-state contract shared by NestJS API, prototype handler, API client, CLI, and local E2E.
servers:
  - url: http://localhost:3000/api
  - url: https://api.raibitserver.app
security:
  - bearerAuth: []
paths:
...
  /projects: { get: { summary: List visible projects, parameters: [ { $ref: '#/components/parameters/limit' }, { $ref: '#/components/parameters/cursor' } ], responses: { '200': { description: Keyset-paginated projects } } }, post: { summary: Create project using subject organization, responses: { '201': { description: Created } } } }
...
components:
  parameters:
...
    limit: { name: limit, in: query, required: false, description: Maximum rows per page, schema: { type: integer, minimum: 1, maximum: 1000, default: 200 } }
    cursor: { name: cursor, in: query, required: false, description: Opaque versioned keyset cursor returned by the previous page, schema: { type: string, maxLength: 1024 } }
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...